Use size_t for image sizes.
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2515 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/loader/DOS4GWBinaryFile.h
@@ -210,7 +210,7 @@ private:
     LXHeader* m_pLXHeader; // Pointer to lx header
     LXObject* m_pLXObjects; // Pointer to lx objects
     LXPage* m_pLXPages; // Pointer to lx pages
-    int m_cbImage; // Size of image
+    size_t m_cbImage; // Size of image
     //int		m_cReloc;				// Number of relocation entries
     //DWord*	m_pRelocTable;			// The relocation table
     unsigned char *base; // Beginning of the loaded image
